+++ src/Course.php
@@ -2,11 +2,12 @@
 
 namespace Drupal\anu_lms;
 
-use Drupal\Core\Entity\EntityRepositoryInterface;
-use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\node\NodeInterface;
 use Drupal\Core\Url;
+use Drupal\Core\Entity\EntityRepositoryInterface;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\File\FileUrlGeneratorInterface;
+use Drupal\Core\Entity\EntityStorageInterface;
 
 /**
  * Methods that operate with course nodes.
@@ -18,21 +19,28 @@ class Course {
    *
    * @var \Drupal\Core\Entity\EntityStorageInterface
    */
-  protected $nodeStorage;
+  protected EntityStorageInterface $nodeStorage;
+
+  /**
+   * The paragraph storage.
+   *
+   * @var \Drupal\Core\Entity\EntityStorageInterface
+   */
+  protected EntityStorageInterface $paragraphStorage;
 
   /**
    * The file url generator.
    *
    * @var \Drupal\Core\File\FileUrlGeneratorInterface
    */
-  protected $fileUrlGenerator;
+  protected FileUrlGeneratorInterface $fileUrlGenerator;
 
   /**
    * The entity repository service.
    *
    * @var \Drupal\Core\Entity\EntityRepositoryInterface
    */
-  protected $entityRepository;
+  protected EntityRepositoryInterface $entityRepository;
 
   /**
    * Constructs service.
@@ -46,6 +54,7 @@ class Course {
    */
   public function __construct(EntityTypeManagerInterface $entity_type_manager, FileUrlGeneratorInterface $file_url_generator, EntityRepositoryInterface $entity_repository) {
     $this->nodeStorage = $entity_type_manager->getStorage('node');
+    $this->paragraphStorage = $entity_type_manager->getStorage('paragraph');
     $this->fileUrlGenerator = $file_url_generator;
     $this->entityRepository = $entity_repository;
   }
@@ -59,26 +68,38 @@ class Course {
    * @return \Drupal\node\NodeInterface|bool
    *   Lesson or Quiz node object.
    */
-  public function getFirstAccessibleLesson(NodeInterface $course) {
-    $modules = $course->get('field_course_module')->referencedEntities();
+  public function getFirstAccessibleLesson(NodeInterface $course): NodeInterface {
+    $modules = $course->get('field_course_module')->getValue();
     foreach ($modules as $module) {
+      $module_paragraph = $this->paragraphStorage->load($module['target_id']);
+
+      // Make sure that the paragraph was successfully loaded.
+      if (empty($module_paragraph)) {
+        continue;
+      }
 
       /** @var \Drupal\node\NodeInterface[] $lessons */
-      $lessons = $module->field_module_lessons->referencedEntities();
+      $lessons = $module_paragraph->get('field_module_lessons')->getValue();
       foreach ($lessons as $lesson) {
-        if ($lesson->access('view')) {
-          return $lesson;
+        /** @var \Drupal\node\NodeInterface $lesson_node */
+        $lesson_node = $this->nodeStorage->load($lesson['target_id']);
+        if ($lesson_node && $lesson_node->access('view')) {
+          return $lesson_node;
         }
       }
 
-      if (!$module->field_module_assessment) {
+      // Ensure that the field with assessment exists, given that it's a part of
+      // submodule anu_lms_assessments.
+      if (!$module_paragraph->hasField('field_module_assessment')) {
         continue;
       }
       /** @var \Drupal\node\NodeInterface[] $quizzes */
-      $quizzes = $module->field_module_assessment->referencedEntities();
+      $quizzes = $module_paragraph->get('field_module_assessment')->getValue();
       foreach ($quizzes as $quiz) {
-        if ($quiz->access('view')) {
-          return $quiz;
+        /** @var \Drupal\node\NodeInterface $quiz_node */
+        $quiz_node = $this->nodeStorage->load($quiz['target_id']);
+        if ($quiz_node && $quiz_node->access('view')) {
+          return $quiz_node;
         }
       }
     }
@@ -92,34 +113,35 @@ class Course {
    * @param \Drupal\node\NodeInterface $course
    *   Course node object.
    *
-   * @return \Drupal\node\NodeInterface[]
+   * @return array
    *   Flat list of lessons and quizzes.
    */
-  public function getLessonsAndQuizzes(NodeInterface $course) {
+  public function getLessonsAndQuizzes(NodeInterface $course): array {
     $nodes = &drupal_static('anu_lms_course_lessons', []);
     if (!empty($nodes[$course->id()])) {
       return $nodes[$course->id()];
     }
 
+    // Initialize list of lessons and quizzes for the current course.
+    $nodes[$course->id()] = [];
+
     // Get course modules.
+    /** @var \Drupal\paragraphs\ParagraphInterface[] $modules */
     $modules = $course->get('field_course_module')->referencedEntities();
     foreach ($modules as $module) {
-
-      // Get module's lessons.
-      /** @var \Drupal\node\NodeInterface[] $lessons */
-      $lessons = $module->field_module_lessons->referencedEntities();
-      foreach ($lessons as $lesson) {
-        $nodes[$course->id()][] = $lesson;
+      $lesson_values = $module->get('field_module_lessons')->getValue();
+      foreach ($lesson_values as $lesson_value) {
+        $nodes[$course->id()][] = $lesson_value['target_id'];
       }
 
-      // Get module's quiz.
-      if (!$module->field_module_assessment) {
+      // Ensure anu_lms_assessments module is enabled.
+      if (!$module->hasField('field_module_assessment')) {
         continue;
       }
-      /** @var \Drupal\node\NodeInterface[] $quizzes */
-      $quizzes = $module->field_module_assessment->referencedEntities();
-      foreach ($quizzes as $quiz) {
-        $nodes[$course->id()][] = $quiz;
+
+      $quiz_values = $module->get('field_module_assessment')->getValue();
+      foreach ($quiz_values as $quiz_value) {
+        $nodes[$course->id()][] = $quiz_value['target_id'];
       }
     }
 
@@ -148,7 +170,7 @@ class Course {
    * @return \Drupal\Core\Url
    *   Url object for redirect.
    */
-  public function getFinishRedirectUrl(NodeInterface $course) {
+  public function getFinishRedirectUrl(NodeInterface $course): Url {
     $uri = $course->field_course_finish_button->uri;
     if (!empty($uri)) {
       return Url::fromUri($uri);
@@ -162,11 +184,11 @@ class Course {
    * @param \Drupal\node\NodeInterface $course
    *   Course node object.
    *
-   * @return \Drupal\Core\Url
-   *   Url object for redirect.
+   * @return string
+   *   Finish button label.
    */
-  public function getFinishText(NodeInterface $course) {
-    return $course->field_course_finish_button->title;
+  public function getFinishText(NodeInterface $course): string {
+    return $course->field_course_finish_button->title ?? '';
   }
 
   /**
@@ -178,7 +200,7 @@ class Course {
    * @return int
    *   Number of quizzes and lessons.
    */
-  public function countLessons(NodeInterface $course) {
+  public function countLessons(NodeInterface $course): int {
     return count($this->getLessons($course));
   }
 
@@ -191,7 +213,7 @@ class Course {
    * @return int
    *   Number of quizzes.
    */
-  public function countQuizzes(NodeInterface $course) {
+  public function countQuizzes(NodeInterface $course): int {
     return count($this->getQuizzes($course));
   }
 
@@ -204,7 +226,7 @@ class Course {
    * @return \Drupal\node\NodeInterface[]
    *   Lessons in the course.
    */
-  public function getLessons(NodeInterface $course) {
+  public function getLessons(NodeInterface $course): array {
     $modules = $course->get('field_course_module')->referencedEntities();
     $lessonsInCourse = [];
 
@@ -230,7 +252,7 @@ class Course {
    * @return \Drupal\node\NodeInterface[]
    *   Quizzes in the course.
    */
-  public function getQuizzes(NodeInterface $course) {
+  public function getQuizzes(NodeInterface $course): array {
     $modules = $course->get('field_course_module')->referencedEntities();
     $quizzesInCourse = [];
 
@@ -259,7 +281,7 @@ class Course {
    * @return array
    *   List of urls.
    */
-  public function getLessonsAndQuizzesUrls(NodeInterface $course) {
+  public function getLessonsAndQuizzesUrls(NodeInterface $course): array {
     $modules = $course->get('field_course_module')->referencedEntities();
     $urls = [];
 
@@ -301,7 +323,7 @@ class Course {
    * @return string[]
    *   Audio URLs.
    */
-  public function getAudios(NodeInterface $course) {
+  public function getAudios(NodeInterface $course): array {
     $lessons = $this->getLessons($course);
 
     $audiosInCourse = [];