+++ src/CoursesPage.php
@@ -2,12 +2,12 @@
 
 namespace Drupal\anu_lms;
 
-use Drupal\Core\Entity\EntityInterface;
+use Drupal\Core\Cache\Cache;
+use Drupal\Core\Cache\CacheableMetadata;
+use Drupal\Core\Cache\CacheBackendInterface;
+use Drupal\node\NodeInterface;
+use Drupal\Core\Entity\EntityStorageInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
-use Drupal\Core\Language\LanguageManagerInterface;
-use Drupal\anu_lms\Event\CoursesPageDataGeneratedEvent;
-use Drupal\node\Entity\Node;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
  * Normalizer service for the given entity.
@@ -19,158 +19,46 @@ class CoursesPage {
    *
    * @var \Drupal\Core\Entity\EntityStorageInterface
    */
-  protected $nodeStorage;
+  protected EntityStorageInterface $nodeStorage;
 
   /**
-   * The taxonomy storage.
+   * The paragraph storage.
    *
-   * @var \Drupal\taxonomy\TermStorageInterface
+   * @var \Drupal\Core\Entity\EntityStorageInterface
    */
-  protected $taxonomyStorage;
+  protected EntityStorageInterface $paragraphStorage;
 
   /**
-   * The language manager.
+   * Cache service.
    *
-   * @var \Drupal\Core\Language\LanguageManagerInterface
+   * @var \Drupal\Core\Cache\CacheBackendInterface
    */
-  protected $languageManager;
-
-  /**
-   * The event dispatcher.
-   *
-   * @var \Symfony\Component\EventDispatcher\EventDispatcherInterface
-   */
-  protected $dispatcher;
-
-  /**
-   * The normalizer.
-   *
-   * @var \Drupal\anu_lms\Normalizer
-   */
-  protected $normalizer;
-
-  /**
-   * The course page service.
-   *
-   * @var \Drupal\anu_lms\Course
-   */
-  protected $course;
+  protected CacheBackendInterface $cache;
 
   /**
    * Constructs service.
    *
    * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
    *   The entity type manager.
-   * @param \Drupal\Core\Language\LanguageManagerInterface $language_manager
-   *   The language manager.
-   * @param \Symfony\Component\EventDispatcher\EventDispatcherInterface $dispatcher
-   *   The event dispatcher.
-   * @param \Drupal\anu_lms\Normalizer $normalizer
-   *   The normalizer.
-   * @param \Drupal\anu_lms\Course $course
-   *   The Course service.
+   * @param \Drupal\Core\Cache\CacheBackendInterface $cache
+   *   Cache service.
    */
-  public function __construct(EntityTypeManagerInterface $entity_type_manager, LanguageManagerInterface $language_manager, EventDispatcherInterface $dispatcher, Normalizer $normalizer, Course $course) {
+  public function __construct(EntityTypeManagerInterface $entity_type_manager, CacheBackendInterface $cache) {
     $this->nodeStorage = $entity_type_manager->getStorage('node');
-    $this->taxonomyStorage = $entity_type_manager->getStorage('taxonomy_term');
-    $this->languageManager = $language_manager;
-    $this->dispatcher = $dispatcher;
-    $this->normalizer = $normalizer;
-    $this->course = $course;
-  }
-
-  /**
-   * Returns normalized Courses page data.
-   *
-   * @param \Drupal\Core\Entity\EntityInterface $node
-   *   Courses page node.
-   *
-   * @return array
-   *   An array containing current node and list of referenced courses.
-   */
-  public function getCoursesPageData(EntityInterface $node) {
-    $referenced_entities = $node->get('field_courses_content')->referencedEntities();
-
-    // Get all selected categories from attached paragraphs.
-    $project_category_ids = [];
-    /** @var \Drupal\Core\Entity\EntityInterface[] $referenced_entities */
-    foreach ($referenced_entities as $referenced_entity) {
-      if ($referenced_entity->bundle() === 'course_category') {
-        if (!$referenced_entity->get('field_course_category')->isEmpty()) {
-          $project_category_ids[] = intval($referenced_entity->get('field_course_category')->first()->getString());
-        }
-      }
-    }
-    $project_category_ids = array_unique($project_category_ids);
-
-    // Get list of referenced Course entities.
-    $courses = $this->getCoursesByCategories($project_category_ids);
-    $normalized_courses = [];
-    foreach ($courses as $course) {
-      $normalized_course = $this->normalizer->normalizeEntity($course, [
-        'max_depth' => 1,
-        // Pass the categories requested as context so additional logic
-        // can be performed like the course being part of a sequence within
-        // a category.
-        // @see \Drupal\anu_lms\CourseProgress
-        'course_page_categories' => $this->taxonomyStorage->loadMultiple($project_category_ids),
-      ]);
-      if (!empty($normalized_course)) {
-        $normalized_courses[] = $normalized_course;
-      }
-    }
-
-    // Get list of Courses pages for every Course.
-    $courses_pages_by_course = [];
-    foreach ($courses as $course) {
-      $courses_pages = $this->getCoursesPagesByCourse($course);
-      $normalized_courses_pages = [];
-      foreach ($courses_pages as $courses_page) {
-        $normalized_courses_pages[] = [
-          'courses_page' => $this->normalizer->normalizeEntity($courses_page, ['max_depth' => 1]),
-        ];
-      }
-
-      $courses_pages_by_course[] = [
-        'course_id' => $course->id(),
-        'courses_pages' => $normalized_courses_pages,
-      ];
-    }
-
-    // Get first accessible Lesson URL.
-    $first_lesson_url_by_course = [];
-    $current_language = $this->languageManager->getCurrentLanguage();
-    foreach ($courses as $course) {
-      $lesson = $this->course->getFirstAccessibleLesson($course);
-      if (!empty($lesson)) {
-        $first_lesson_url_by_course[] = [
-          'course_id' => $course->id(),
-          'first_lesson_url' => $lesson->toUrl('canonical', ['language' => $current_language])->toString(),
-        ];
-      }
-    }
-
-    $pageData = [
-      $node->bundle() => $this->normalizer->normalizeEntity($node, ['max_depth' => 3]),
-      'courses' => $normalized_courses,
-      'courses_pages_by_course' => $courses_pages_by_course,
-      'first_lesson_url_by_course' => $first_lesson_url_by_course,
-    ];
-    $event = new CoursesPageDataGeneratedEvent($pageData, $node);
-    $this->dispatcher->dispatch(CoursesPageDataGeneratedEvent::EVENT_NAME, $event);
-    return $event->getPageData();
+    $this->paragraphStorage = $entity_type_manager->getStorage('paragraph');
+    $this->cache = $cache;
   }
 
   /**
    * Returns list of Course entities by given category ids.
    *
    * @param array $category_ids
-   *   List of category ids (int).
+   *   List of category ids.
    *
-   * @return array|EntityInterface[]
+   * @return array|NodeInterface[]
    *   An array of Course entities.
    */
-  public function getCoursesByCategories(array $category_ids) {
+  public function getCoursesByCategories(array $category_ids): array {
     if (empty($category_ids)) {
       return [];
     }
@@ -188,42 +76,104 @@ class CoursesPage {
   /**
    * Returns list of Courses page entities by given Course.
    *
-   * @param \Drupal\Core\Entity\EntityInterface $course
+   * @param \Drupal\node\NodeInterface $course
    *   Course entity.
    *
-   * @return array|EntityInterface[]
+   * @return array|\Drupal\node\NodeInterface[]
    *   An array of Courses page entities.
    */
-  public function getCoursesPagesByCourse(EntityInterface $course) {
+  public function getCoursesPagesByCourse(NodeInterface $course): array {
     if (empty($course)) {
       return [];
     }
 
     // Gathering course categories.
     $category_ids = array_column($course->get('field_course_category')->getValue(), 'target_id');
-
     if (empty($category_ids)) {
       return [];
     }
 
-    // Gathering courses page's categories parargaphs.
-    $courses_page_categories = \Drupal::entityQuery('paragraph')
-      ->condition('type', 'course_category')
-      ->condition('field_course_category', $category_ids, 'IN')
-      ->execute();
+    // Given that this method can be used for every course on a courses page,
+    // we cache mapping between courses pages content type and the categories
+    // represented by their paragraphs.
+    $cache_id = 'anu_lms:courses_pages_by_course_map';
+    $courses_pages_mapping = &drupal_static($cache_id, FALSE);
 
-    if (empty($courses_page_categories)) {
-      return [];
+    // If not found in static cache - lookup in the backend cache.
+    if ($courses_pages_mapping === FALSE) {
+      $cache = $this->cache->get($cache_id);
+      if (!empty($cache->data)) {
+        $courses_pages_mapping = $cache->data;
+      }
     }
 
-    // Gathering courses page IDs.
-    $courses_page_categories = array_values($courses_page_categories);
-    $courses_page_ids = \Drupal::entityQuery('node')
-      ->condition('type', 'courses_page')
-      ->condition('field_courses_content', $courses_page_categories, 'IN')
-      ->execute();
+    // If mapping was not found in cache, then we need to build it from
+    // scratch.
+    if ($courses_pages_mapping === FALSE) {
+      $courses_pages_mapping = [];
+      $cacheable_metadata = new CacheableMetadata();
 
-    return !empty($courses_page_ids) ? Node::loadMultiple($courses_page_ids) : [];
+      // Load all courses_pages (we expected have just several pages max per
+      // site, so it should be OK).
+      $courses_pages = $this->nodeStorage->loadByProperties(['type' => 'courses_page']);
+      foreach ($courses_pages as $courses_page) {
+        $course_categories_paragraphs = $courses_page->get('field_courses_content')->referencedEntities();
+        foreach ($course_categories_paragraphs as $course_categories_paragraph) {
+          $course_category_tid = $course_categories_paragraph->get('field_course_category')->getString();
+          $courses_pages_mapping[$course_category_tid][] = $courses_page->id();
+          $cacheable_metadata->addCacheableDependency($courses_page);
+        }
+      }
+
+      // Add mapping to backend cache with cache tags to trigger recalculation
+      // only when content of the courses cache changes.
+      $this->cache->set(
+        $cache_id,
+        $courses_pages_mapping,
+        Cache::PERMANENT,
+        $cacheable_metadata->getCacheTags()
+      );
+    }
+
+    $courses_page_ids = [];
+    foreach ($category_ids as $category_id) {
+      if (!empty($courses_pages_mapping[$category_id])) {
+        $courses_page_ids += $courses_pages_mapping[$category_id];
+      }
+    }
+
+    return !empty($courses_page_ids) ? $this->nodeStorage->loadMultiple($courses_page_ids) : [];
+  }
+
+  /**
+   * Load list of courses page URLs for each course.
+   *
+   * Note that this is needed only for offline mode.
+   *
+   * @param array $courses
+   *   List of course nodes.
+   *
+   * @return array
+   *   Course page URLs per course.
+   *
+   * @throws \Drupal\Core\Entity\EntityMalformedException
+   */
+  public function getCoursesPageUrlsByCourse(array $courses): array {
+    $courses_page_urls_by_course = [];
+    foreach ($courses as $course) {
+      $courses_pages = $this->getCoursesPagesByCourse($course);
+      $courses_page_urls = [];
+      foreach ($courses_pages as $courses_page_node) {
+        $courses_page_urls[] = $courses_page_node->toUrl()->toString();
+      }
+
+      $courses_page_urls_by_course[] = [
+        'course_id' => (int) $course->id(),
+        'courses_page_urls' => $courses_page_urls,
+      ];
+    }
+
+    return $courses_page_urls_by_course;
   }
 
 }