+++ src/Event/LessonCompletedEvent.php
@@ -3,8 +3,7 @@
 namespace Drupal\anu_lms\Event;
 
 use Drupal\Core\Session\AccountInterface;
-use Symfony\Component\EventDispatcher\Event;
-use Drupal\Core\Entity\EntityInterface;
+use Symfony\Contracts\EventDispatcher\Event;
 
 /**
  * Event that is fired when the Lesson is completed by the user.
@@ -18,26 +17,26 @@ class LessonCompletedEvent extends Event {
    *
    * @var \Drupal\Core\Session\AccountInterface
    */
-  protected $account;
+  protected AccountInterface $account;
 
   /**
-   * The lesson node.
+   * The lesson node ID.
    *
-   * @var \Drupal\Core\Entity\EntityInterface
+   * @var int
    */
-  protected $lesson;
+  protected int $lessonId;
 
   /**
    * Constructor.
    *
    * @param \Drupal\Core\Session\AccountInterface $account
    *   The account.
-   * @param \Drupal\Core\Entity\EntityInterface $lesson
-   *   The lesson.
+   * @param int $lesson_id
+   *   The lesson ID.
    */
-  public function __construct(AccountInterface $account, EntityInterface $lesson) {
+  public function __construct(AccountInterface $account, int $lesson_id) {
     $this->account = $account;
-    $this->lesson = $lesson;
+    $this->lessonId = $lesson_id;
   }
 
   /**
@@ -46,18 +45,18 @@ class LessonCompletedEvent extends Event {
    * @return \Drupal\Core\Session\AccountInterface
    *   The user.
    */
-  public function getAccount() {
+  public function getAccount(): AccountInterface {
     return $this->account;
   }
 
   /**
    * Gets the lesson.
    *
-   * @return \Drupal\Core\Entity\EntityInterface
-   *   The lesson.
+   * @return int
+   *   The lesson ID.
    */
-  public function getLesson() {
-    return $this->lesson;
+  public function getLessonId(): int {
+    return $this->lessonId;
   }
 
 }