+++ src/Plugin/rest/resource/Progress.php
@@ -43,7 +43,7 @@ class Progress extends ResourceBase {
    *
    * @var \Drupal\Core\Session\AccountInterface
    */
-  protected $currentUser;
+  protected AccountInterface $currentUser;
 
   /**
    * The Lesson helper.
@@ -136,7 +136,7 @@ class Progress extends ResourceBase {
       $nodes = $this->nodeStorage->loadMultiple(array_filter($data));
       foreach ($nodes as $node) {
         if (in_array($node->bundle(), ['module_lesson', 'module_assessment'])) {
-          $this->lesson->setCompleted($node);
+          $this->lesson->setCompleted($node->id());
           $results[] = $node->id();
         }
       }