+++ src/Settings.php
@@ -4,6 +4,7 @@ namespace Drupal\anu_lms;
 
 use Drupal\Core\Entity\EntityRepositoryInterface;
 use Drupal\Core\Extension\ModuleExtensionList;
+use Drupal\Core\Extension\ModuleHandlerInterface;
 use Drupal\Core\Language\LanguageManagerInterface;
 use Drupal\Core\Path\PathMatcherInterface;
 use Drupal\config_pages\Entity\ConfigPages;
@@ -19,35 +20,42 @@ class Settings {
    *
    * @var \Drupal\Core\Entity\EntityRepositoryInterface
    */
-  protected $entityRepository;
+  protected EntityRepositoryInterface $entityRepository;
 
   /**
    * The serializer.
    *
    * @var \Symfony\Component\Serializer\Serializer
    */
-  protected $serializer;
+  protected Serializer $serializer;
 
   /**
    * The language manager.
    *
    * @var \Drupal\Core\Language\LanguageManagerInterface
    */
-  protected $languageManager;
+  protected LanguageManagerInterface $languageManager;
 
   /**
    * The path matcher.
    *
    * @var \Drupal\Core\Path\PathMatcherInterface
    */
-  protected $pathMatcher;
+  protected PathMatcherInterface $pathMatcher;
 
   /**
    * The module extension list.
    *
    * @var \Drupal\Core\Extension\ModuleExtensionList
    */
-  protected $moduleExtensionList;
+  protected ModuleExtensionList $moduleExtensionList;
+
+  /**
+   * The module handler service.
+   *
+   * @var \Drupal\Core\Extension\ModuleHandlerInterface
+   */
+  protected ModuleHandlerInterface $moduleHandler;
 
   /**
    * Creates an Settings object.
@@ -62,13 +70,16 @@ class Settings {
    *   The path matcher.
    * @param \Drupal\Core\Extension\ModuleExtensionList $extension_list_module
    *   The module extension list.
+   * @param \Drupal\Core\Extension\ModuleHandlerInterface $module_handler
+   *   The module handler service.
    */
-  public function __construct(EntityRepositoryInterface $entity_repository, Serializer $serializer, LanguageManagerInterface $language_manager, PathMatcherInterface $path_matcher, ModuleExtensionList $extension_list_module) {
+  public function __construct(EntityRepositoryInterface $entity_repository, Serializer $serializer, LanguageManagerInterface $language_manager, PathMatcherInterface $path_matcher, ModuleExtensionList $extension_list_module, ModuleHandlerInterface $module_handler) {
     $this->entityRepository = $entity_repository;
     $this->serializer = $serializer;
     $this->languageManager = $language_manager;
     $this->pathMatcher = $path_matcher;
     $this->moduleExtensionList = $extension_list_module;
+    $this->moduleHandler = $module_handler;
   }
 
   /**
@@ -105,12 +116,22 @@ class Settings {
   }
 
   /**
+   * Returns boolean whether the offline mode can be supported.
+   *
+   * @return bool
+   *   Boolean indicating whether the offline mode is supported.
+   */
+  public function isOfflineSupported(): bool {
+    return $this->moduleHandler->moduleExists('pwa');
+  }
+
+  /**
    * Returns Pwa settings.
    *
    * Code from /pwa/src/Controller/PWAController.php:151
    */
-  public function getPwaSettings() {
-    if (!\Drupal::moduleHandler()->moduleExists('pwa')) {
+  public function getPwaSettings(): ?array {
+    if (!$this->isOfflineSupported()) {
       return NULL;
     }